# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/vaccines/plan/{pet_id}": {
            "get": {
                "description": "Get the vaccination plan of given pet_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vaccine"
                ],
                "summary": "Get vaccination plan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pet id to get vaccines",
                        "name": "pet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "applied",
                            "pending"
                        ],
                        "type": "string",
                        "description": "desired formant for the output",
                        "name": "output",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.VaccinationPlan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            }
        },
        "/vaccines/vaccine": {
            "post": {
                "description": "Create a vaccine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vaccine"
                ],
                "summary": "Creates a Vaccine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "request from telegram",
                        "name": "X-Telegram-App",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chat id of the telegram user",
                        "name": "X-Telegram-Id",
                        "in": "header"
                    },
                    {
                        "description": "vaccine info",
                        "name": "vaccine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Vaccine"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Vaccine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            }
        },
        "/vaccines/vaccine/{id}": {
            "get": {
                "description": "Get vaccine info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vaccine"
                ],
                "summary": "Get a Vaccine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "request from telegram",
                        "name": "X-Telegram-App",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chat id of the telegram user",
                        "name": "X-Telegram-Id",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "id of the vaccine",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Vaccine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit vaccine info given a pet ID and vaccine info needed to update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vaccine"
                ],
                "summary": "Edit a Vaccine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "request from telegram",
                        "name": "X-Telegram-App",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chat id of the telegram user",
                        "name": "X-Telegram-Id",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "id of the vaccine",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "vaccine info",
                        "name": "vaccine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Vaccine"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Vaccine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Vaccine given a pet ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vaccine"
                ],
                "summary": "Delete a Vaccine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "request from telegram",
                        "name": "X-Telegram-App",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chat id of the telegram user",
                        "name": "X-Telegram-Id",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "id of the vaccine",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.APIError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "controller.Vaccine": {
            "type": "object",
            "properties": {
                "animal": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AnimalType"
                        }
                    ],
                    "example": "dog"
                },
                "description": {
                    "type": "string",
                    "example": "vaccine to preventing rage"
                },
                "name": {
                    "type": "string",
                    "example": "anti-rabies"
                },
                "scheduled": {
                    "type": "integer",
                    "example": 365
                }
            }
        },
...
        "model.AnimalType": {
            "type": "string",
            "enum": [
                "dog",
                "cat",
                "bird",
                "hamster"
            ],
            "x-enum-varnames": [
                "Dog",
                "Cat",
                "Bird",
                "Hamster"
            ]
        },
...
        "model.VaccinationPlan": {
            "type": "object",
            "properties": {
                "applied": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Vaccine"
                    }
                },
                "name": {
                    "type": "string"
                },
                "ownerID": {
                    "type": "string"
                },
                "pending": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Vaccine"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.Vaccine": {
            "type": "object",
            "properties": {
                "animal": {
                    "$ref": "#/definitions/model.AnimalType"
                },
                "applied_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "scheduled": {
                    "type": "integer"
                }
            }
        },
...
    }
}